refactor image prompt generation
