+++ assets/data/image_prompts.js
@@ -1,5 +1,11 @@
-export const BASE_IMAGE_PROMPT_TEMPLATE = "A hyper-detailed fantasy Anime-style full-body portrait of a handsome {sex} {race} dressed in tantalizing, tasteful garb in front of a neutral gray background. Reference facial features from the most attractive {sexPlural}. Human proportionality for height is about 7.5 heads tall. Hair: {hair}. Skin: {skin} with light freckles. Eyes: {eyes}. Clothing is vibrant, complementary colors contrasting hair and skin tones with accents, trim, and accessories. Dynamic composition, ultra high definition, delicate details. No hat. No weapon.";
-export const ADDON_IMAGE_PROMPT_TEMPLATE = "Picture Theme: {themeName} — {themeDesc} — {colors}.";
+import { themeColors, getThemeColors } from "./theme_colors.js";
+import { getThemeDescription } from "./theme_descriptions.js";
+import { getRaceColors } from "./race_colors.js";
+
+export const BASE_IMAGE_PROMPT_TEMPLATE =
+  "A hyper-detailed fantasy Anime-style full-body portrait of a handsome {sex} {race} dressed in tantalizing, tasteful garb in front of a neutral gray background. Reference facial features from the most attractive {sexPlural}. Human proportionality for height is about 7.5 heads tall. Hair: {hair}. Skin: {skin} with light freckles. Eyes: {eyes}. Clothing is vibrant, complementary colors contrasting hair and skin tones with accents, trim, and accessories. Dynamic composition, ultra high definition, delicate details. No hat. No weapon.";
+export const ADDON_IMAGE_PROMPT_TEMPLATE =
+  "Picture Theme: {themeName} — {themeDesc} — {colors}.";
 
 export function getRacePrompt(race) {
   switch (race) {
@@ -36,4 +42,32 @@ export function buildImagePrompt({ sex, sexPlural, race, hair, skin, eyes, theme
     .replace('{themeDesc}', themeDesc)
     .replace('{colors}', colors);
   return racePart ? `${base} ${racePart}. ${addon}` : `${base} ${addon}`;
+}
+
+export function composeImagePrompt(character) {
+  const themeIndex = themeColors.find(t => t.name === character.theme)?.index;
+  const pictureTheme = getThemeColors(themeIndex);
+  const descriptor = getThemeDescription(character.theme);
+  const raceCombo = themeIndex ? getRaceColors(character.race, themeIndex) : null;
+
+  const skinColor = character.skinColor || raceCombo?.skin;
+  const skin = skinColor || '';
+  const hair = character.hairColor || raceCombo?.hair || 'brown';
+  const eyes = character.eyeColor || raceCombo?.eyes || 'brown';
+  const sexPlural = character.sex === 'Male' ? 'men' : 'women';
+  const themeName = character.theme || '';
+  const themeDesc = descriptor || '';
+  const colors = pictureTheme.join(', ');
+
+  return buildImagePrompt({
+    sex: character.sex,
+    sexPlural,
+    race: character.race,
+    hair,
+    skin,
+    eyes,
+    themeName,
+    themeDesc,
+    colors,
+  });
 }
